Improves instantiation of Kafka
- Improves instanciation of kafka based on suggested comment
- Updates the unit tests
