+++ tests/test_backend_handler.py
@@ -3,7 +3,7 @@ import json
 import unittest
 from unittest.mock import Mock, MagicMock, patch, call
 
-from DeviceManager.BackendHandler import KafkaHandler
+from DeviceManager.BackendHandler import KafkaHandler, KafkaInstanceHandler
 
 
 class TestBackendHandler(unittest.TestCase):
@@ -39,3 +39,8 @@ class TestBackendHandler(unittest.TestCase):
                   'id': 1, 'created': '2019-08-29T18:18:07.801602+00:00'}
 
         KafkaHandler().configure(device, meta={"service": 'admin'})
+
+    def test_verify_intance_kafka(self):
+        with patch('DeviceManager.BackendHandler.KafkaHandler') as mock_kafka_instance_wrapper:
+            mock_kafka_instance_wrapper.return_value = Mock()
+            self.assertIsNotNone(KafkaInstanceHandler().getInstance(None))
